+++ components/script/dom/canvas/canvas_context.rs
@@ -6,8 +6,8 @@
 
 use euclid::default::Size2D;
 use pixels::Snapshot;
-use script_bindings::canvas::CanvasContext;
 use script_bindings::root::{Dom, DomRoot};
+use script_foo::canvas::CanvasContext;
 use webrender_api::ImageKey;
 
 use crate::DomTypeHolder;
