+++ components/script/dom/canvas/imagebitmaprenderingcontext.rs
@@ -8,9 +8,9 @@ use dom_struct::dom_struct;
 use euclid::default::Size2D;
 use paint_api::SerializableImageData;
 use pixels::Snapshot;
-use script_bindings::canvas::CanvasContext;
 use script_bindings::cell::DomRefCell;
 use script_bindings::reflector::{Reflector, reflect_dom_object};
+use script_foo::canvas::CanvasContext;
 use servo_base::Epoch;
 use webrender_api::units::DeviceIntSize;
 use webrender_api::{ImageDescriptor, ImageFormat, ImageKey};
