+++ components/script/dom/canvas/offscreencanvas.rs
@@ -12,11 +12,12 @@ use js::realm::CurrentRealm;
 use js::rust::{HandleObject, HandleValue};
 use pixels::{EncodedImageType, Snapshot};
 use rustc_hash::FxHashMap;
-use script_bindings::canvas::{CanvasContext, CanvasSizeTrait};
+use script_bindings::canvas::CanvasSizeTrait;
 use script_bindings::cell::{DomRefCell, Ref};
 use script_bindings::inheritance::Castable;
 use script_bindings::reflector::{DomObject, reflect_dom_object_with_proto_and_cx};
 use script_bindings::weakref::WeakRef;
+use script_foo::canvas::CanvasContext;
 use servo_base::id::{OffscreenCanvasId, OffscreenCanvasIndex};
 use servo_canvas_traits::webgl::{GLContextAttributes, WebGLVersion};
 use servo_constellation_traits::{BlobImpl, TransferableOffscreenCanvas};
