+++ components/script/dom/html/htmlcanvaselement.rs
@@ -13,10 +13,11 @@ use js::rust::{HandleObject, HandleValue};
 use layout_api::HTMLCanvasData;
 use pixels::{EncodedImageType, Snapshot};
 use rustc_hash::FxHashMap;
-use script_bindings::canvas::{CanvasContext, CanvasSizeTrait};
+use script_bindings::canvas::CanvasSizeTrait;
 use script_bindings::cell::{DomRefCell, Ref};
 use script_bindings::reflector::DomObject;
 use script_bindings::weakref::WeakRef;
+use script_foo::canvas::CanvasContext;
 use servo_base::Epoch;
 use servo_canvas_traits::webgl::{GLContextAttributes, WebGLVersion};
 use servo_constellation_traits::BlobImpl;
