+++ components/script/dom/webxr/xrwebgllayer.rs
@@ -7,8 +7,8 @@ use std::convert::TryInto;
 use dom_struct::dom_struct;
 use euclid::{Rect, Size2D};
 use js::rust::HandleObject;
-use script_bindings::canvas::CanvasContext;
 use script_bindings::reflector::reflect_dom_object_with_proto;
+use script_foo::canvas::CanvasContext;
 use servo_canvas_traits::webgl::{WebGLCommand, WebGLContextId, WebGLTextureId};
 use webxr_api::{ContextId as WebXRContextId, LayerId, LayerInit, Viewport};
