+++ components/script_foo/Cargo.toml
@@ -1,0 +1,70 @@
+[package]
+name = "servo-script-foo"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+publish.workspace = true
+rust-version.workspace = true
+repository.workspace = true
+description.workspace = true
+# BSD-3-Clause comes from third_party/ply, which doesn't end up in any produced build-artifact,
+# but is used during the build via build.rs.
+license = "MPL-2.0 AND BSD-3-Clause"
+
+
+[lib]
+name = "script_foo"
+path = "lib.rs"
+
+[dependencies]
+bitflags = { workspace = true }
+crossbeam-channel = { workspace = true }
+deny_public_fields = { workspace = true }
+dom_struct = { workspace = true }
+encoding_rs = { workspace = true }
+html5ever = { workspace = true }
+indexmap = { workspace = true }
+js = { workspace = true }
+jstraceable_derive = { workspace = true }
+keyboard-types = { workspace = true }
+libc = { workspace = true }
+log = { workspace = true }
+malloc_size_of = { workspace = true }
+malloc_size_of_derive = { workspace = true }
+num-traits = { workspace = true }
+parking_lot = { workspace = true }
+phf = { workspace = true }
+profile_traits = { workspace = true }
+pixels = { workspace = true }
+euclid = { workspace = true }
+regex = { workspace = true }
+servo-base = { workspace = true }
+servo-config = { workspace = true }
+servo-url = { workspace = true }
+servo_arc = { workspace = true }
+script_bindings = { workspace = true }
+smallvec = { workspace = true }
+stylo = { workspace = true }
+stylo_atoms = { workspace = true }
+tendril = { workspace = true }
+tracing = { workspace = true, optional = true }
+webxr-api = { workspace = true, optional = true }
+xml5ever = { workspace = true }
+zeroize = { workspace = true, features = ["zeroize_derive"] }
+
+[build-dependencies]
+phf_codegen = { workspace = true }
+phf_shared = { workspace = true }
+serde_json = { workspace = true }
+
+[features]
+bluetooth = []
+gamepad = []
+testbinding = []
+tracing = ["dep:tracing"]
+webgpu = []
+webxr = ["webxr-api"]
+
+[lints.rust]
+unexpected_cfgs = { level = "warn", check-cfg = ['cfg(crown)'] }
+unsafe_op_in_unsafe_fn = { level = "allow" }
